+++ core/shared/src/main/scala/zio/logging/FilteredLogger.scala
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2019-2023 John A. De Goes and the ZIO Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package zio.logging
+
+import zio.{ Cause, FiberId, FiberRefs, LogLevel, LogSpan, Trace, ZLogger }
+
+final case class FilteredLogger[Message, Output](logger: zio.ZLogger[Message, Output], filter: LogFilter[Message])
+    extends ZLogger[Message, Option[Output]] {
+
+  override def apply(
+    trace: Trace,
+    fiberId: FiberId,
+    logLevel: LogLevel,
+    message: () => Message,
+    cause: Cause[Any],
+    context: FiberRefs,
+    spans: List[LogSpan],
+    annotations: Map[String, String]
+  ): Option[Output] =
+    if (filter(trace, fiberId, logLevel, message, cause, context, spans, annotations)) {
+      Some(logger(trace, fiberId, logLevel, message, cause, context, spans, annotations))
+    } else None
+
+  def withFilter(newFilter: LogFilter[Message]): FilteredLogger[Message, Output] = FilteredLogger(logger, newFilter)
+}
